pyproject.toml
[tool.poetry]
name = "wedding-backend"
version = "0.1.0"
description = "Backend FastAPI server for the wedding site"
authors = ["Caluminus <caluminus@hotmail.co.uk>"]
license = "MIT"

[tool.poetry.scripts]
start = "uvicorn src.main:app --reload"


[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
strawberry-graphql = {extras = ["fastapi"], version = "^0.102.2"}
tortoise-orm = "^0.18.1"
asyncpg = "^0.25.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
